backend/requirements.txt
# Core Framework
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0

# Database
sqlalchemy>=2.0.35,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0

# Data Validation
pydantic>=2.9.0,<3.0.0

# Security & Authentication
python-multipart>=0.0.18
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
bcrypt==4.0.1

# HTTP Client (AI proxy)
httpx>=0.27.0,<1.0.0

# Utilities
python-dotenv>=1.0.1,<2.0.0
